requirements.txt
boto3
s3fs
numpy
pandas
torch
pydub
pyarrow
apache-airflow
ffmpeg-python